Making a fix to GroupFile test suite
The problem - When GroupFile test suite would run, it would add a user to a group, upload a file and everything, but didn't make the user leave the group after it was all done. This eventually maxed out the group for anyone else to join.
The Fix - Add LeaveGroup method into the suite so it cleans itself up properly after each test.